+++ lib/nft.mli
@@ -16,7 +16,9 @@ type run_error = {
 }
 
 val list_ruleset_invocation : unit -> invocation
+val apply_invocation : string -> invocation
 val run : invocation -> (string, run_error) result
 val list_ruleset_with_runner : (invocation -> (string, run_error) result) -> (string, run_error) result
 val list_ruleset : unit -> (string, run_error) result
+val apply : string -> (unit, run_error) result
 val string_of_run_error : run_error -> string